NIT Move the BSP-related `JMH` tests to the `JmhTests` suite
